🚚 Move all Spanner transaction files to the same folder
