+++ src/transaction/index.ts
@@ -1,4 +1,2 @@
 export * from './firestore-pubsub/index.js';
 export * from './spanner-outbox/index.js';
-export * from './spanner-readonly-transaction.js';
-export { SpannerStateTransaction } from './spanner-state-transaction.js';
