+++ src/transaction/spanner-outbox/index.ts
@@ -1,5 +1,7 @@
 export { SpannerOutboxEvent } from './event.js';
 export { SpannerOutboxTransactionModule } from './module.js';
+export * from './readonly-transaction.js';
 export { SpannerOutboxTransactionRunner } from './runner.js';
 export { SpannerOutboxSender } from './sender.js';
+export { SpannerStateTransaction } from './state-transaction.js';
 export * from './transaction.js';
